requirements.txt
streamlit>=1.28.0
huggingface-hub>=0.20.0
requests>=2.31.0
python-dotenv>=1.0.0
pandas>=2.0.0
plotly>=5.17.0
json5>=0.12.1